Fix invalid operation on shader related to the tint being the wrong type

+++ src/main/resources/assets/tails/shader/threetint_frag.glsl
@@ -3,7 +3,7 @@
 varying vec2 TexCoord;
 
 uniform sampler2D tex;
-uniform vec3[3] tints; // todo encode rgb into one int value again to save memory?
+uniform ivec3[3] tints; // todo encode rgb into one int value again to save memory?
 
 void main() {
   vec4 texCol = texture2D(tex, TexCoord);
